requirements.txt
streamlit
supabase
requests
cloudscraper
trafilatura
playwright
google-generativeai
python-dotenv
